+++ internal/cmd/gtrace/main.go
@@ -67,11 +67,15 @@ func main() {
 		// We should respect Go suffixes like `_linux.go`.
 		name, tags, ext := splitOSArchTags(&buildCtx, gofile)
 		openFile := func(name string) (*os.File, func()) {
-			p := filepath.Join(workDir, name)
 			var f *os.File
 			// nolint: gofumpt
 			// nolint: nolintlint
-			f, err = os.OpenFile(p, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
+			// nolint: gosec
+			f, err = os.OpenFile(
+				filepath.Join(workDir, filepath.Clean(name)),
+				os.O_WRONLY|os.O_CREATE|os.O_TRUNC,
+				0600,
+			)
 			if err != nil {
 				log.Fatal(err)
 			}