+++ CHANGELOG.md
@@ -1,7 +1,6 @@
-* Added address to error description which thrown from `conn.take()`
-* Renamed package `internal/db` to `internal/database` for exclude collisions with variable name `db`
-* Added `conn.Ping(ctx)` method for check avalability of `conn.Conn`
+* Added `conn.Ping(ctx)` method for check availability of `conn.Conn`
 * Refactored `cluster.Cluster.Get(ctx)` for return only available connection (instead return any connection from balancer)
+* Added address to error description which thrown from `conn.take()`
 * Renamed package `internal/db` to `internal/database` for exclude collisions with variable name `db`
 
 ## v3.18.0